refactor: use subscription callback instead of referenceAreaSourceAtom
